Fix #221: Allow to define robots.txt file per domain
The implementation permit to use `/robots-files/*.txt` files as
substitution for `robots.txt` on a used domain basis.

If no file exists corresponding to domain, then the `default.txt`
file is served.

The downside of this implementation is that there is a duplicated
file for geokrety.org and www.geokrety.org domains :/ Does using
a symlink between thos two files will work when checkout in M$ os?