Crude fix for the `?waypoint` query param not being loaded properly
That's the problem I described previously in the email thread titled "Small API suggestion improvement" (yeah, there's a typo in the subject).

I suspect this should fix it, but I have not verified it (I just edited the code in the repo, I did not run it)! If you could spare some time to verify this fix please, and pull it to prod, I think some users would be grateful!