fix: Allow "met" moves at the same datetime
If the GK is marked as "non-collectible" then we allow to use the same datetime as another move.
Useful for the use case from samuel: Just imagine one GK present at a geocaching.com mega-event, on a car for example. You will have thousands of people wanting to log it in a few hours.

If the GK is not marked as non-collectible, the "met" move type will virtually change the GK position, in this case, the move date should be strict.

Should fix the issue on gk side for cgeo/cgeo#5182